Finally have boolean intersects working
Radius is converting to a polygon now and comparable to other weather alert polygons. Need to transfer code that converts circles to polygon to another .ts page and make it look neater.

Next step is saving the values so the radius doesn't lose its shape everytime the app closes. Also use APNs to send notifications to apple phones.